refactor($model): change time basis
- add msPerTick
