+++ src/constants/defaults.js
@@ -6,9 +6,10 @@ export const midiMessageGeneratorOptions = {
   nowCb: () => window.performance.now()
 };
 
-export const bpm = 120;
+const bpm = 120;
 /** @type {number} ticks per beat `ticks = minutes * bpm * tpb` */
 export const ticksPerBeat = 480;
 export const beatsPerBar = 4;
+export const msPerTick = 60 * 1e3 / bpm / ticksPerBeat;
 /** 8 bars */
 export const songDuration = 8 * ticksPerBeat * beatsPerBar;
