+++ src/renderer/interactors/Song/index.js
@@ -1,0 +1,6 @@
+import { SONG_ID } from '../../../constants/ids';
+import { Song } from '../../store/models';
+
+export const getSongData = ():Song => {
+  return Song.query().whereId(SONG_ID).first();
+};
