+++ src/types/index.js
@@ -1,7 +1,8 @@
 // @flow
 // eslint-disable-next-line no-unused-vars
 import * as models from '../renderer/store/models';
-
+import type { Ms, Tick, Sec, Magnitude, SamplingRate, Frequency, NoteNumber } from './units';
+export type { Ms, Tick, Sec, Magnitude, SamplingRate, Frequency, NoteNumber };
 /**
  * @typedef Spectrogram
  * @property {Array<Array<Number>>} magnitude2d
@@ -61,26 +62,3 @@ export type Now = () => number;
 export type NoteAction = models.NoteOn | models.Modulation | models.NoteOff;
 
 export type Callable = (...args: Array<any>) => any;
-/**
- * Unit is Hz
- * eg 48000
- */
-export type SamplingRate = number;
-
-/**
- * Unit is Hz
- * eg: 440
- */
-export type Frequency = number;
-
-/**
- * Unit is MIDI Note Number.
- * from 0 to 127
- * eg: 60
- */
-export type NoteNumber = number;
-
-/**
- * from 0.0 to 1.0
- */
-export type Magnitude = number;
