+++ src/types/units.js
@@ -1,0 +1,26 @@
+export type Tick = number;
+export type Ms = number;
+export type Sec = number;
+/**
+ * Unit is Hz
+ * eg 48000
+ */
+export type SamplingRate = number;
+
+/**
+ * Unit is Hz
+ * eg: 440
+ */
+export type Frequency = number;
+
+/**
+ * Unit is MIDI Note Number.
+ * from 0 to 127
+ * eg: 60
+ */
+export type NoteNumber = number;
+
+/**
+ * from 0.0 to 1.0
+ */
+export type Magnitude = number;
